feat: implement TabLink component for project navigation and update button styles
